backend/requirements.txt
# backend/requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MongoDB (Primary Database)
motor==3.3.2
pymongo==4.5.0
beanie==1.23.0  # Optional: ODM for MongoDB

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# API & Validation
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0

# Async Support
aiofiles==23.2.1
httpx==0.25.1
aiohttp==3.9.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.32
ta==0.11.0  # Technical Analysis

# Machine Learning & AI
scikit-learn==1.3.2
transformers==4.35.2  # For FinBERT
torch==2.1.1
tensorflow==2.15.0  # Optional

# NLP & Sentiment Analysis
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Scheduling & Background Tasks
apscheduler==3.10.4
celery==5.3.4  # Optional for distributed tasks
redis==5.0.1  # For caching and Celery broker

# Visualization & Charting
plotly==5.18.0
matplotlib==3.8.2

# News & Data APIs
newsapi-python==0.2.7
alpha-vantage==2.3.1
finnhub-python==2.4.19

# Notifications
twilio==8.10.1  # For WhatsApp/SMS
sendgrid==6.11.0  # For email

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0  # Optional for metrics
sentry-sdk==1.38.0  # Optional for error tracking

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1  # For testing HTTP clients
mongomock-motor==0.0.21  # For testing MongoDB

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# CORS & Middleware
fastapi-limiter==0.1.5
slowapi==0.1.9  # Rate limiting

# Environment Management
python-decouple==3.8

# Remove these (SQL-related):
# sqlalchemy
# alembic
# psycopg2-binary
# sqlite3 (built-in, no need to list)